+++ Assets/Scripts/BasicClasses/Route.cs
@@ -10,7 +10,7 @@ public enum RouteType
 public class Route
 {
     public Node destination;
-    public int distance;    // km
+    public float distance;    // km
     public RouteType routeType;
     public float beauty;    // 0-1
